+++ src/application/components/Layouts/screens.ts
@@ -1,0 +1,6 @@
+export enum Screens {
+  Cache = "cache",
+  Queries = "queries",
+  Mutations = "mutations",
+  Explorer = "explorer",
+}
