+++ src/application/components/Mutations/__tests__/Mutations.test.tsx
@@ -3,22 +3,27 @@ import { within, waitFor } from "@testing-library/react";
 import user from "@testing-library/user-event";
 
 import { renderWithApolloClient } from "../../../utilities/testing/renderWithApolloClient";
-import { client, GET_MUTATIONS } from "../../../index";
+import { client, currentClient, GET_MUTATIONS } from "../../../index";
 import { Mutations } from "../Mutations";
 
 jest.mock("../MutationViewer", () => ({
   MutationViewer: () => null,
 }));
 
+
+const clientId = "client-1";
+
 describe("<Mutations />", () => {
   const mutations = [
     {
       id: 0,
+      clientId,
       __typename: "Mutation",
       name: null,
     },
     {
       id: 1,
+      clientId,
       __typename: "Mutation",
       name: "AddColorToFavorites",
     },
@@ -34,14 +39,22 @@ describe("<Mutations />", () => {
   });
 
   test("queries render in the sidebar", async () => {
+    currentClient(clientId);
     client.writeQuery({
       query: GET_MUTATIONS,
       data: {
-        mutationLog: {
-          mutations,
-          count: mutations.length,
-        },
+        client: {
+          id: clientId,
+          __typename: "Client",
+          mutationLog: {
+            mutations,
+            count: mutations.length,
+          },
+        }
       },
+      variables: {
+        clientId
+      }
     });
 
     const { getByTestId } = renderWithApolloClient(
@@ -63,14 +76,22 @@ describe("<Mutations />", () => {
   });
 
   test("renders query name", async () => {
+    currentClient(clientId)
     client.writeQuery({
       query: GET_MUTATIONS,
       data: {
-        mutationLog: {
-          mutations,
-          count: mutations.length,
-        },
+        client: {
+          id: clientId,
+          __typename: "Client",
+          mutationLog: {
+            mutations,
+            count: mutations.length,
+          }
+        }
       },
+      variables: {
+        clientId
+      }
     });
 
     const { getByTestId } = renderWithApolloClient(
